+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - (breaking) Remove use of ->error, ->info, and ->validate since donut.system has removed them. Replace with clojure.tools.logging and exceptions.
   - Add `clojure.tools.logging` as a direct dependency.
 - (breaking) Remove salmon.signal/signal!, delete!, early-validate!, start!, and start!. They are no longer needed due to the removal of ->error, ->info, and ->validate.
+- Fix an issue where the `:lint?` stack option was ignored.
 
 ## v0.14.0 (2024-01-16)